Remember that in Swift, enums can implement protocols just like structs and classes. So, we can create a FlickrEndpoint entity that implements the protocol endpoint we defined earlier. Now, for  every new endpoint we want to support, we can just introduce a new case to the enum which will force us to handle it in the code below helping us prevent unintended states. Additionally, it doesn’t always mean we need to write a lot of new code because we can rely on the default case for an enum like we do for the baseURL. We can only make changes to the part of the endpoint requests that change and let everything else defer to the default implementation.
Another point to mention is how we might store our API Key if the project requires it. I think for a take home project, it’s totally fine just to include it in the project in plaintext, but you should mention either through a comment or in the interview, that in the real-world implementation you’d prefer a much more rigorous authentication flow and that including the API key in the client is inherently unsecure.