Let’s go ahead and define some models. As you can see they all implement the Codable protocol and are derived from looking at the JSON the Flickr API sends back. Generally, you’d probably want to split these out into their respective files, but for the video I’ll leave it like this.
A few things to note here, we’re making them structs because we want our models to be immutable. We don’t want to accidentally change their values elsewhere in the code without realizing it.

The reason we’re creating these Codable models here, is our interaction with our API endpoints will likely be sending or receiving some JSON object, so we can think about that communication as “transacting” with Codable objects which means our network request calls can be generic functions operating on Codables